Developed spheroid-fitter Made summary_data real

+++ GUI/libspheroid.h
@@ -12,7 +12,7 @@ void simulate_step(int *);
 void terminate_run(int *);
 void get_dimensions(int *, int *, int *, int *, double *, int *, int *, bool *, double *, double *);
 void get_scene(int *, int *);
-void get_summary(int *, int *, int *);
+void get_summary(double *, int *, int *);
 void get_concdata(int *, int *, double *, double *);
 void get_ic_concdata(int *, int *, double *, double *);
 //void get_distdata(int *, DIST_SET *);
